App: extract command line options to separate module
